o Classes I forgot to commit o Changed Money's amount from BigDecimal to Long.
git-svn-id: https://scm.ops4j.org/repos/ops4j/projects/qi4j/apps/chronos@8237 9b982a3c-3ae5-0310-a4bc-d9a3335569bd

+++ core/src/main/java/org/qi4j/chronos/model/mixins/MoneyDescriptorMixin.java
@@ -34,7 +34,7 @@ public final class MoneyDescriptorMixin implements Descriptor
             currencyCode = Character.SPACE_SEPARATOR + currency.getCurrencyCode();
         }
 
-        BigDecimal amount = money.getAmount();
+        long amount = money.getAmount();
         return amount + currencyCode;
     }
